refactor: UrlStatisticResponse (#26)
